Added CUberData and Memory

+++ PS1Ubr/SUberHeader.cs
@@ -12,19 +12,21 @@ namespace PS1Ubr
     /// </summary>
     public struct SUberHeader
     {
-        internal UInt16[] FourCC;
-        internal uint VersionA;
-        internal uint VersionB;
-        internal uint EntryCount;
-        internal uint Vec3Count;
-        internal uint Vec2Count;
-        internal uint LookupCount;
-        internal uint U32Count;
-        internal uint MeshDataSize;
-        internal uint Field_28;
+        internal Byte[] FourCC;
+        internal UInt32 VersionA;
+        internal UInt32 VersionB;
+        internal UInt32 EntryCount;
+        internal UInt32 Vec3Count;
+        internal UInt32 Vec2Count;
+        internal UInt32 LookupCount;
+        internal UInt32 U32Count;
+        internal UInt32 MeshDataSize;
+        internal UInt32 Field_28;
 
-        internal static readonly UInt16[] FourCCValue = new UInt16[4] { 'u', 'b', 'e', 'r' };
-        internal static readonly uint VersionAValue = 1;
-        internal static readonly uint VersionBValue = 1;
+        internal static readonly Byte[] FourCCValue = 
+            new Byte[4] { (byte)('u'), (byte)('b'), (byte)('e'), (byte)('r') };
+
+        internal static readonly UInt32 VersionAValue = 1;
+        internal static readonly UInt32 VersionBValue = 1;
     }
 }
